+++ cookbook/use_cases/apps/sql/app.py
@@ -4,6 +4,14 @@ import streamlit as st
 from agents import get_sql_agent
 from agno.agent import Agent, RunResponse
 from agno.utils.log import logger
+from utils import (
+    CUSTOM_CSS,
+    add_message,
+    display_tool_calls,
+    export_chat_history,
+    load_data_and_knowledge,
+    restart_agent,
+)
 
 # Page configuration
 st.set_page_config(
@@ -13,159 +21,15 @@ st.set_page_config(
     initial_sidebar_state="expanded",
 )
 
-# Custom CSS with dark mode support
-st.markdown(
-    """
-    <style>
-    /* Main Styles */
-    .main-title {
-        text-align: center;
-        background: linear-gradient(45deg, #FF4B2B, #FF416C);
-        -webkit-background-clip: text;
-        -webkit-text-fill-color: transparent;
-        font-size: 3em;
-        font-weight: bold;
-        padding: 1em 0;
-    }
-    .subtitle {
-        text-align: center;
-        color: #666;
-        margin-bottom: 2em;
-    }
-    .stButton button {
-        width: 100%;
-        border-radius: 20px;
-        margin: 0.2em 0;
-        transition: all 0.3s ease;
-    }
-    .stButton button:hover {
-        transform: translateY(-2px);
-        box-shadow: 0 5px 15px rgba(0,0,0,0.1);
-    }
-    .chat-container {
-        border-radius: 15px;
-        padding: 1em;
-        margin: 1em 0;
-        background-color: #f5f5f5;
-    }
-    .sql-result {
-        background-color: #f8f9fa;
-        border-radius: 10px;
-        padding: 1em;
-        margin: 1em 0;
-        border-left: 4px solid #FF4B2B;
-    }
-    .status-message {
-        padding: 1em;
-        border-radius: 10px;
-        margin: 1em 0;
-    }
-    .success-message {
-        background-color: #d4edda;
-        color: #155724;
-    }
-    .error-message {
-        background-color: #f8d7da;
-        color: #721c24;
-    }
-    /* Dark mode adjustments */
-    @media (prefers-color-scheme: dark) {
-        .chat-container {
-            background-color: #2b2b2b;
-        }
-        .sql-result {
-            background-color: #1e1e1e;
-        }
-    }
-    </style>
-""",
-    unsafe_allow_html=True,
-)
-
-
-def load_data_and_knowledge():
-    """Load F1 data and knowledge base if not already done"""
-    from load_f1_data import load_f1_data
-    from load_knowledge import load_knowledge
-
-    if "data_loaded" not in st.session_state:
-        with st.spinner("🔄 Loading F1 data into database..."):
-            load_f1_data()
-        with st.spinner("📚 Loading knowledge base..."):
-            load_knowledge()
-        st.session_state["data_loaded"] = True
-        st.success("✅ F1 data and knowledge loaded successfully!")
-
-
-def add_message(role: str, content: str) -> None:
-    """Safely add a message to the session state"""
-    if "messages" not in st.session_state or not isinstance(
-        st.session_state["messages"], list
-    ):
-        st.session_state["messages"] = []
-    st.session_state["messages"].append({"role": role, "content": content})
-
-
-def restart_agent():
-    """Reset the agent and clear chat history"""
-    logger.debug("---*--- Restarting agent ---*---")
-    st.session_state["sql_agent"] = None
-    st.session_state["messages"] = []
-    st.rerun()
-
-
-def export_chat_history():
-    """Export chat history as markdown"""
-    if "messages" in st.session_state:
-        chat_text = "# F1 SQL Agent - Chat History\n\n"
-        for msg in st.session_state["messages"]:
-            role = "🤖 Assistant" if msg["role"] == "agent" else "👤 User"
-            chat_text += f"### {role}\n{msg['content']}\n\n"
-        return chat_text
-    return ""
-
-
-def display_tool_calls(tool_calls_container, tools):
-    """Display tool calls in a streamlit container with expandable sections.
-
-    Args:
-        tool_calls_container: Streamlit container to display the tool calls
-        tools: List of tool call dictionaries containing name, args, content, and metrics
-    """
-    with tool_calls_container.container():
-        for tool_call in tools:
-            _tool_name = tool_call.get("tool_name")
-            _tool_args = tool_call.get("tool_args")
-            _content = tool_call.get("content")
-            _metrics = tool_call.get("metrics")
-
-            with st.expander(
-                f"🛠️ {_tool_name.replace('_', ' ').title()}", expanded=False
-            ):
-                if isinstance(_tool_args, dict) and "query" in _tool_args:
-                    st.code(_tool_args["query"], language="sql")
-
-                if _tool_args and _tool_args != {"query": None}:
-                    st.markdown("**Arguments:**")
-                    st.json(_tool_args)
-
-                if _content:
-                    st.markdown("**Results:**")
-                    try:
-                        st.json(_content)
-                    except Exception as e:
-                        st.markdown(_content)
-
-                if _metrics:
-                    st.markdown("**Metrics:**")
-                    st.json(_metrics)
+# Load custom CSS with dark mode support
+st.markdown(CUSTOM_CSS, unsafe_allow_html=True)
 
 
 def main() -> None:
     # Header
     st.markdown("<h1 class='main-title'>F1 SQL Agent</h1>", unsafe_allow_html=True)
     st.markdown(
-        "<p class='subtitle'>Your intelligent F1 data companion powered by Agno</p>",
+        "<p class='subtitle'>Your intelligent F1 data analyst powered by Agno</p>",
         unsafe_allow_html=True,
     )
 
@@ -245,69 +109,56 @@ def main() -> None:
     if "messages" not in st.session_state or not isinstance(
         st.session_state["messages"], list
     ):
-        st.session_state["messages"] = [
-            {
-                "role": "assistant",
-                "content": "👋 Ask me about F1 data from 1950 to 2020!",
-            }
-        ]
+        st.session_state["messages"] = []
 
-    # Main chat interface
-    chat_container = st.container()
-    with chat_container:
-        # Clear the container before displaying messages
-        st.empty()
+    # Get user input
+    if prompt := st.chat_input("👋 Ask me about F1 data from 1950 to 2020!"):
+        add_message("user", prompt)
 
-        # Display chat history
-        for message in st.session_state["messages"]:
-            if message["role"] == "system":
-                continue
-            with st.chat_message(message["role"]):
-                st.markdown(message["content"])
+    # Display chat history
+    for message in st.session_state["messages"]:
+        if message["role"] == "system":
+            continue
+        with st.chat_message(message["role"]):
+            st.markdown(message["content"])
 
-        # Generate response for last user message
-        last_message = (
-            st.session_state["messages"][-1] if st.session_state["messages"] else None
-        )
-        if last_message and last_message.get("role") == "user":
-            question = last_message["content"]
-            with st.chat_message("assistant"):
-                resp_container = st.empty()
-                with st.spinner("🤔 Thinking..."):
-                    response = ""
-                    try:
-                        # Create container for tool calls
-                        tool_calls_container = st.empty()
+    # Generate response for last user message
+    last_message = (
+        st.session_state["messages"][-1] if st.session_state["messages"] else None
+    )
+    if last_message and last_message.get("role") == "user":
+        question = last_message["content"]
+        with st.chat_message("assistant"):
+            resp_container = st.empty()
+            with st.spinner("🤔 Thinking..."):
+                response = ""
+                try:
+                    # Create container for tool calls
+                    tool_calls_container = st.empty()
 
-                        # Run the agent and stream the response
-                        run_stream: Iterator[RunResponse] = sql_agent.run(
-                            question, stream=True
-                        )
-                        for _resp_chunk in run_stream:
-                            # Display tool calls if available
-                            if _resp_chunk.tools and len(_resp_chunk.tools) > 0:
-                                display_tool_calls(
-                                    tool_calls_container, _resp_chunk.tools
-                                )
+                    # Run the agent and stream the response
+                    run_stream: Iterator[RunResponse] = sql_agent.run(
+                        question, stream=True
+                    )
+                    for _resp_chunk in run_stream:
+                        # Display tool calls if available
+                        if _resp_chunk.tools and len(_resp_chunk.tools) > 0:
+                            display_tool_calls(tool_calls_container, _resp_chunk.tools)
 
-                            # Display response
-                            if _resp_chunk.content is not None:
-                                response += _resp_chunk.content
-                                resp_container.markdown(response)
+                        # Display response
+                        if _resp_chunk.content is not None:
+                            response += _resp_chunk.content
+                            resp_container.markdown(response)
 
-                        st.session_state["messages"].append(
-                            {"role": "assistant", "content": response}
-                        )
-                    except Exception as e:
-                        error_message = f"Sorry, I encountered an error: {str(e)}"
-                        st.error(error_message)
-                        st.session_state["messages"].append(
-                            {"role": "assistant", "content": error_message}
-                        )
-
-        # Get user input
-        if prompt := st.chat_input("Ask about F1 data..."):
-            add_message("user", prompt)
+                    st.session_state["messages"].append(
+                        {"role": "assistant", "content": response}
+                    )
+                except Exception as e:
+                    error_message = f"Sorry, I encountered an error: {str(e)}"
+                    st.error(error_message)
+                    st.session_state["messages"].append(
+                        {"role": "assistant", "content": error_message}
+                    )
 
 
 if __name__ == "__main__":
